requirements.txt
discord.py
openai
scikit-learn
numpy